Add remove chain account logic (#7)
* Add remove chain account logic

* Add unittest

* update unittest

* update hasLogin to getUAVersion

+++ src/main/java/com/webank/wecross/account/service/db/LoginTokenTableJPA.java
@@ -2,6 +2,6 @@ package com.webank.wecross.account.service.db;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoginTokenTableJPA extends JpaRepository<LoginTokenTableBean, Long> {
+public interface LoginTokenTableJPA extends JpaRepository<LoginTokenTableBean, Integer> {
     LoginTokenTableBean findByToken(String token);
 }
